refactor: MVC 디자인 패턴 적용
