+++ src/main/java/calculator/view/InputView.java
@@ -1,8 +1,8 @@
-package calculator;
+package calculator.view;
 
 import camp.nextstep.edu.missionutils.Console;
 
-public class InputHandler {
+public class InputView {
     private static final String INPUT_GUIDE_MESSAGE = "덧셈할 문자열을 입력해 주세요.";
 
     public static String inputString() {
