+++ src/test/java/calculator/ApplicationTest.java
@@ -4,6 +4,7 @@ import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import calculator.constant.ErrorMessage;
 import camp.nextstep.edu.missionutils.test.NsTest;
 import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
